+++ lib/cuckoo/common/cape_utils.py
@@ -95,6 +95,7 @@ except ImportError as e:
 
 processing_conf = Config("processing")
 
+HAVE_FLARE_CAPA = False
 if processing_conf.flare_capa.enabled:
     try:
         import capa.main
@@ -117,8 +118,15 @@ if processing_conf.flare_capa.enabled:
     except ImportError:
         HAVE_FLARE_CAPA = False
         print("FLARE-CAPA missed, pip3 install flare-capa")
-else:
-    HAVE_FLARE_CAPA = False
+
+
+HAVE_VBA2GRAPH = False
+if processing_conf.vba2graph.enabled:
+    try:
+        from lib.cuckoo.common.office.vba2graph import vba2graph_from_vba_object, vba2graph_gen
+        HAVE_VBA2GRAPH = True
+    except ImportError:
+        HAVE_VBA2GRAPH = False
 
 
 suppress_parsing_list = ["Cerber", "Emotet_Payload", "Ursnif", "QakBot"]
@@ -320,6 +328,17 @@ def flare_capa_details(file_path: str, category: str, on_demand: bool=False) -> 
 
     return capa_dictionary
 
+def vba2graph_func(file_path: str, id: str, on_demand: bool=False):
+    if HAVE_VBA2GRAPH and processing_conf.vba2graph.enabled and (processing_conf.vba2graph.on_demand is False or on_demand is True:
+        try:
+            vba2graph_path = os.path.join(CUCKOO_ROOT, "storage", "analyses", str(id), "vba2graph")
+            if not os.path.exists(vba2graph_path):
+                os.makedirs(vba2graph_path)
+            vba_code = vba2graph_from_vba_object(file_path)
+            if vba_code:
+                vba2graph_gen(vba_code, vba2graph_path)
+        except Exception as e:
+            log.info(e)
 
 def hash_file(method, path):
     """Calculates an hash on a file by path.